fix: Retry PipelineRuns that fail due to a Tekton race condition
If we see a particular string in the failure message for a
`PipelineRun`, we know it's due to a race condition in Tekton where
the `PipelineRun` starts getting handled by the controller but its
cache of `Pipeline`s doesn't yet contain the relevant `Pipeline`. So
when we see that, let's reset the `ProwJob` and delete the
existing-but-failed `PipelineRun`(s) so that we can retry.

fixes https://github.com/jenkins-x/jx/issues/5324

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>